# swagger-template.yaml
openapi: 3.0.3
info:
  title: 智养云·智慧养老OA平台 API
  version: 1.0.0
  description: |
    后端全量接口 Swagger 风格模板（可直接补充字段与示例）。
    统一响应：{ code, message, data }
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/family/register:
    post:
      tags: [Family]
      summary: 家属注册
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FamilyRegister' }
      responses:
        '200': { description: OK }
  /api/family/bindElder:
    post:
      tags: [Family]
      summary: 绑定老人
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FamilyBind' }
      responses:
        '200': { description: OK }
...
  /api/admin/disease:
    get:
      tags: [Disease]
      summary: 疾病列表
      responses:
        '200': { description: OK }
    post:
      tags: [Disease]
      summary: 创建疾病
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Disease' }
      responses:
        '200': { description: OK }
  /api/admin/disease/{id}:
    put:
      tags: [Disease]
      summary: 更新疾病
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Disease' }
      responses:
        '200': { description: OK }
    delete:
      tags: [Disease]
      summary: 删除疾病
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: OK }
...
  /api/admin/disease/{id}/forbidden-tags:
    get:
      tags: [Disease]
      summary: 查询禁忌标签
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: OK }
    put:
      tags: [Disease]
      summary: 保存禁忌标签
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tagIds:
                  type: array
                  items: { type: integer }
      responses:
        '200': { description: OK }
...
  /api/store/order/preview:
    post:
      tags: [Store]
      summary: 下单预览
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreOrderPreview' }
      responses:
        '200': { description: OK }
  /api/store/order/submit:
    post:
      tags: [Store]
      summary: 下单提交
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StoreOrderSubmit' }
      responses:
        '200': { description: OK }
  /api/store/order/page:
    get:
      tags: [Store]
      summary: 订单分页
      parameters: [{ $ref: '#/components/parameters/PageNo' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { description: OK }
  /api/store/order/{id}:
    get:
      tags: [Store]
      summary: 订单详情
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200': { description: OK }
...
  /api/inventory/page:
    get:
      tags: [Inventory]
      summary: 库存分页
      parameters: [{ $ref: '#/components/parameters/PageNo' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { description: OK }
...
  /api/inventory/adjust:
    post:
      tags: [Inventory]
      summary: 库存盘点调整
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InventoryAdjust' }
      responses:
        '200': { description: OK }
  /api/bill/page:
    get:
      tags: [Billing]
      summary: 账单分页
      parameters: [{ $ref: '#/components/parameters/PageNo' }, { $ref: '#/components/parameters/PageSize' }, { name: month, in: query, schema: { type: string } }]
      responses:
        '200': { description: OK }
  /api/bill/{elderId}:
    get:
      tags: [Billing]
      summary: 老人账单
      parameters: [{ $ref: '#/components/parameters/ElderIdPath' }, { name: month, in: query, schema: { type: string } }]
      responses:
        '200': { description: OK }
  /api/bill/generate:
    post:
      tags: [Billing]
      summary: 生成账单
      parameters: [{ name: month, in: query, schema: { type: string } }]
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 10 }
...
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
    ElderIdPath:
      name: elderId
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    FamilyRegister:
      type: object
      properties:
        phone: { type: string }
        code: { type: string }
    FamilyBind:
      type: object
      properties:
        familyUserId: { type: integer }
        elderId: { type: integer }
        relation: { type: string }
...
    Disease:
      type: object
      properties:
        id: { type: integer }
        diseaseName: { type: string }
...
    StoreOrderPreview:
      type: object
      properties:
        elderId: { type: integer }
        productId: { type: integer }
        qty: { type: integer }
    StoreOrderSubmit:
      type: object
      properties:
        elderId: { type: integer }
        productId: { type: integer }
        qty: { type: integer }
...
    InventoryAdjust:
      type: object
      properties:
        productId: { type: integer }
        adjustQty: { type: integer }
        reason: { type: string }
...